refactor : 리뷰사항 반영 - 추상화관련 수정
