+++ src/main/java/nextstep/subway/auth/application/UserDetailService.java
@@ -1,7 +1,5 @@
 package nextstep.subway.auth.application;
 
-import nextstep.subway.member.domain.LoginMember;
-
 public interface UserDetailService {
-    public LoginMember loadUserByUsername(String email);
+    UserDetails loadUserByUsername(String email);
 }
