+++ src/main/java/nextstep/subway/auth/application/UserDetails.java
@@ -1,0 +1,5 @@
+package nextstep.subway.auth.application;
+
+public interface UserDetails {
+    boolean checkPassword(String password);
+}
