+++ src/main/java/nextstep/subway/auth/ui/authentication/SessionAuthenticationInterceptor.java
@@ -4,7 +4,6 @@ import nextstep.subway.auth.application.UserDetailService;
 import nextstep.subway.auth.domain.Authentication;
 import nextstep.subway.auth.infrastructure.SecurityContext;
 import nextstep.subway.auth.ui.converter.AuthenticationConverter;
-import nextstep.subway.member.application.CustomUserDetailsService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
