+++ src/main/java/nextstep/subway/auth/ui/persistence/SecurityContextInterceptor.java
@@ -1,5 +1,6 @@
 package nextstep.subway.auth.ui.persistence;
 
+import nextstep.subway.auth.infrastructure.SecurityContext;
 import nextstep.subway.auth.infrastructure.SecurityContextHolder;
 import org.springframework.web.servlet.HandlerInterceptor;
 
@@ -8,10 +9,22 @@ import javax.servlet.http.HttpServletResponse;
 
 public abstract class SecurityContextInterceptor implements HandlerInterceptor {
 
-    /*
-    * preHandle
-    * SecurityContextHolder에 인증정보 담기
-    * */
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        if (SecurityContextHolder.getContext().getAuthentication() != null) {
+            return true;
+        }
+
+        SecurityContext securityContext = getSecurityContext(request);
+
+        if (securityContext != null) {
+            SecurityContextHolder.setContext(securityContext);
+        }
+        return true;
+    }
+
+    abstract SecurityContext getSecurityContext(HttpServletRequest request);
+
     @Override
     public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
         SecurityContextHolder.clearContext();
